pyproject.toml
[tool.poetry]
name = "diplomwork"
version = "0.1.0"
description = ""
authors = ["Evgeny Egorov <eaegorov42@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.1.3"
python-dotenv = "^1.0.1"
rest-framework-simplejwt = "^0.0.2"
drf-yasg = "^1.21.8"
django-celery-beat = "^2.7.0"
psycopg2-binary = "^2.9.10"
coverage = "^7.6.8"
requests = "^2.32.3"
redis = "^5.2.0"
celery = "^5.4.0"
pytest = "^8.3.3"
flake8 = "^7.1.1"
djangorestframework-simplejwt = "^5.3.1"
djangorestframework-swagger = "^0.2.2"
django-cors-headers = "^4.6.0"
pillow = "^11.0.0"
black = "^24.10.0"
django-filter = "^24.3"
django-autoslug = "^1.9.9"
django-uuid-primary-key = "^1.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
